+++ packages/config/src/chains.ts
@@ -30,7 +30,6 @@ import {
   ftmwh,
   glmr,
   hdx,
-  hko,
   ibtc,
   intr,
   kar,
@@ -448,7 +447,12 @@ export const hydration = new Parachain({
   name: 'Hydration',
   parachainId: 2034,
   ss58Format: 63,
-  ws: 'wss://hydradx-rpc.dwellir.com',
+  ws: [
+    'wss://hydradx-rpc.dwellir.com',
+    'wss://rpc.hydradx.cloud',
+    'wss://rpc.helikon.io/hydradx',
+    'wss://hydradx.paras.dotters.network',
+  ],
 });
 
 export const hydrationAlphanet = new Parachain({
@@ -1114,11 +1118,6 @@ export const moonriver = new EvmParachain({
       id: '108457044225666871745333730479173774551',
     },
     {
-      address: '0xffffffFF394054BCDa1902B6A6436840435655a3',
-      asset: hko,
-      id: '76100021443485661246318545281171740067',
-    },
-    {
       address: '0xFfFFFFfF08220AD2E6e157f26eD8bD22A336A0A5',
       asset: kar,
       id: '10810581592933651521121702237638664357',
@@ -1315,28 +1314,6 @@ export const parallel = new Parachain({
   parachainId: 2012,
   ss58Format: 172,
   ws: 'wss://parallel-rpc.dwellir.com',
-});
-
-export const parallelHeiko = new Parachain({
-  assetsData: [
-    {
-      asset: movr,
-      id: 113,
-    },
-    // NOTE: no meta for native token
-    {
-      asset: hko,
-      metadataId: 0,
-    },
-  ],
-  ecosystem: Ecosystem.Kusama,
-  genesisHash:
-    '0x64a1c658a48b2e70a7fb1ad4c39eea35022568c20fc44a6e2e3d0a57aee6053b',
-  key: 'parallel-heiko',
-  name: 'Parallel Heiko',
-  parachainId: 2085,
-  ss58Format: 110,
-  ws: 'wss://heiko-rpc.parallel.fi',
 });
 
 export const peaqAlphanet = new Parachain({
@@ -1866,7 +1843,11 @@ export const zeitgeist = new Parachain({
   parachainId: 2092,
   ss58Format: 73,
   usesChainDecimals: true,
-  ws: 'wss://zeitgeist-rpc.dwellir.com',
+  ws: [
+    'wss://main.rpc.zeitgeist.pm/ws',
+    'wss://zeitgeist-rpc.dwellir.com',
+    'wss://zeitgeist.api.onfinality.io/public-ws',
+  ],
 });
 
 export const chainsList: AnyChain[] = [
@@ -1900,7 +1881,6 @@ export const chainsList: AnyChain[] = [
   nodle,
   originTrailAlphanet,
   parallel,
-  parallelHeiko,
   peaqChain,
   peaqEvm,
   peaqAlphanet,