+++ packages/config/src/configs/index.ts
@@ -32,7 +32,6 @@ import { neurowebConfig } from './neuroweb';
 import { nodleConfig } from './nodle';
 import { originTrailAlphanetConfig } from './originTrailAlphanet';
 import { parallelConfig } from './parallel';
-import { parallelHeikoConfig } from './parallelHeiko';
 import { peaqConfig } from './peaq';
 import { peaqAlphanetConfig } from './peaqAlphanet';
 import { peaqEvmConfig } from './peaqEvm';
@@ -84,7 +83,6 @@ export const chainsConfigList: ChainConfig[] = [
   nodleConfig,
   originTrailAlphanetConfig,
   parallelConfig,
-  parallelHeikoConfig,
   peaqConfig,
   peaqEvmConfig,
   peaqAlphanetConfig,